requirements.txt
groq
streamlit
python-dotenv
PyPDF2
python-docx
fpdf2
fastapi
uvicorn
python-multipart
reportlab
pdf2image